Update documentation for sgate command
I noticed the documentation for the sgate command
was slightly incorrect; in particular, the command
was not actually presented in the code snippet
above the graphic! As such, users might be a bit
confused as to what the command does. We incorporate
an explicit call to the command to clarify that.