Update hyperlinks to CQuIC GitHub pages
This commit updates the README and tutorial to point
people in the right direction as far as the CQuIC
GitHub pages are concerned. Recently, we migrated our
username from CQuIC-GitHub to just CQuIC. Hence the need
to update links and whatnot.

In particular, this commit changes the obvious links
in the README, as well as the link in the abstract of the
tutorial.