Update xyoptions to allow for XeLaTeX compiling
Based on a user-submitted issue, compiling the tutorial
in XeLaTeX resulted in a pdf for which there were no
circuits drawn! This commit fixes this issue by replacing
the calls to some of the xyoptions{} with xyoptions{all},
which was observed to fix the problem. The documentation
has been changed to reflect this modification.