Add parameter to \barrier command for horizontal offset
This commit adds a new parameter to the \barrier command to specify a
manual offset to place the barrier from placement. The distance between
elements on a circuit is not fixed and figure out how to place the
barrier may require manual adjustment.

This also adds documentation on how to use the \barrier command to
Qtutorial.tex so there will be a guide on how to use it.